Improves Timer start/stop usage
This patch removes the need for the final Timer.stop() call, and
instead delegates that responsibility to the Timer.reset() call.
As a result of this, the internal Node managed by Timer defers the
ACTUAL resetting until a subsequent Timer.start() is called; this
allows the calling code to query the Timer properly AFTER it has
been reset (since it's just waiting for another start() before it
clears out any data).