Last functional deque-based child commit
This patch is will be the last before I attempt to remove the
usage of std::deque as the Node::Children container type (as it
fails entirely to compiled in MSVC). Ideally, I'll make Timer
simply recreate the Node it manages after every call to reset(),
then using std::move to swap that previous Node graph to the
caller (which will likely be some version of the Profile object).